Move spec parsed data to common helpers
Move spec parsed data to common helpers
