+++ spec/models/manager_refresh/helpers/spec_parsed_data.rb
