+++ spec/models/manager_refresh/save_inventory/graph_of_inventory_collections_targeted_refresh_spec.rb
@@ -1,5 +1,5 @@
 require_relative 'spec_helper'
-require_relative 'spec_parsed_data'
+require_relative '../helpers/spec_parsed_data'
 require_relative 'init_data_helper'
 
 describe ManagerRefresh::SaveInventory do
